+++ graphics/include/ignition/common/Animation.hh
@@ -34,6 +34,7 @@ namespace ignition
     class KeyFrame;
     class PoseKeyFrame;
     class NumericKeyFrame;
+    class AnimationPrivate;
     class TrajectoryInfoPrivate;
 
     /// \class Animation Animation.hh ignition/common/Animation.hh
@@ -70,6 +71,14 @@ namespace ignition
       /// \brief Return the current time position
       /// \return The time position in seconds
       public: double Time() const;
+
+      /// \brief Return the interpolateX value
+      /// \return The interpolateX value
+      public: bool InterpolateX() const;
+
+      /// \brief Set the interpolateX value
+      /// \param[in] _interpolateX The interpolateX value
+      public: void SetInterpolateX(const bool _interpolateX);
 
       /// \brief Return the number of key frames in the animation
       /// \return The number of keyframes
@@ -114,7 +123,7 @@ namespace ignition
       protected: bool loop;
 
       /// \brief array of keyframe type alias
-      protected: typedef std::vector<common::KeyFrame*> KeyFrame_V;
+      protected: typedef std::vector<common::KeyFrame *> KeyFrame_V;
 
       /// \brief array of key frames
       protected: KeyFrame_V keyFrames;
@@ -131,7 +140,7 @@ namespace ignition
       /// \param[in] _length Length of the animation in seconds
       /// \param[in] _loop True == loop the animation
       public: PoseAnimation(const std::string &_name,
-                            const double _length, const bool _loop);
+                  const double _length, const bool _loop);
 
       /// \brief Destructor
       public: virtual ~PoseAnimation();